Move some models to more logical namespace
