+++ src/Altinn.Profile.Core/User.ProfileSettings/LanguageType.cs
@@ -24,22 +24,5 @@
         /// Sami language. Inactive.
         /// </summary>
         public const string SE = "se";
-
-        /// <summary>
-        /// Gets the language code corresponding to the specified Altinn 2 language code.
-        /// </summary>
-        /// <param name="altinn2Code">The Altinn 2 language code (e.g., 1044 for Norwegian Bokmål, 2068 for Norwegian Nynorsk, 1033 for English).</param>
-        /// <returns>The language code as a string ("nb", "nn", or "en"). Defaults to "nb" if the code is not recognized.</returns>
-        public static string GetFromAltinn2Code(int altinn2Code)
-        {
-            return altinn2Code switch
-            {
-                1044 => NB,
-                2068 => NN,
-                1033 => EN,
-                1083 => SE,
-                _ => NB
-            };
-        }
     }
 }
