+++ src/Altinn.Profile.Core/User/UserProfileService.cs
@@ -6,8 +6,6 @@ using Altinn.Register.Contracts;
 
 using Microsoft.Extensions.Options;
 
-using static Altinn.Register.Contracts.PartyUrn;
-
 namespace Altinn.Profile.Core.User;
 
 /// <summary>
