+++ src/Altinn.Profile.Integrations/ContactRegister/Models/ContactRegisterChangesLog.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Immutable;
 using System.Text.Json.Serialization;
 
-namespace Altinn.Profile.Integrations.ContactRegister;
+namespace Altinn.Profile.Integrations.ContactRegister.Models;
 
 /// <summary>
 /// Represents the changes to a person's contact preferences from the contact register.
